+++ src/sw_stat.f95
@@ -2,16 +2,16 @@ module mod_stat
 
   interface
      function sw_dens0(t,s)
-       REAL*8, INTENT(IN)                        :: t(:),s(:)
-       REAL*8, ALLOCATABLE, DIMENSION (:)        :: sw_dens0
+       REAL, INTENT(IN)                        :: t(:),s(:)       
+       REAL, ALLOCATABLE, DIMENSION (:)        :: sw_dens0
      end function sw_dens0
      function sw_pres(depth,lat)
-       REAL*8, INTENT(IN)                        :: depth(:),lat(:)
-       REAL*8, ALLOCATABLE, DIMENSION (:)        :: sw_pres
+       REAL, INTENT(IN)                        :: depth(:),lat(:)       
+       REAL, ALLOCATABLE, DIMENSION (:)        :: sw_pres
      end function sw_pres
      function sw_seck(t,s,pres)
-       REAL*8, INTENT(IN)                        :: t(:),s(:),pres(:)
-       REAL*8, ALLOCATABLE, DIMENSION (:)        :: sw_seck
+       REAL, INTENT(IN)                        :: t(:),s(:),pres(:)       
+       REAL, ALLOCATABLE, DIMENSION (:)        :: sw_seck
      end function sw_seck
   end interface
 
@@ -21,8 +21,8 @@ contains
     IMPLICIT NONE
     
     INTEGER                                     :: km
-    REAL*8, ALLOCATABLE, DIMENSION(:)             :: t ,s ,dens0 ,pres 
-    REAL*8, ALLOCATABLE, DIMENSION(:)             :: depth ,lat ,seck ,dens
+    REAL, ALLOCATABLE, DIMENSION(:)             :: t ,s ,dens0 ,pres 
+    REAL, ALLOCATABLE, DIMENSION(:)             :: depth ,lat ,seck ,dens
     
     allocate ( dens0(km) ,pres(km) ,seck(km))
     
@@ -41,32 +41,32 @@ contains
     
     IMPLICIT NONE
 
-!    REAL*8, ALLOCATABLE, DIMENSION (:)       :: dens_temp, dens_zero, rho
-!    REAL*8, ALLOCATABLE, DIMENSION (:)       :: t, s
+!    REAL*4, ALLOCATABLE, DIMENSION (:)       :: dens_temp, dens_zero, rho
+!    REAL*4, ALLOCATABLE, DIMENSION (:)       :: t, s
     INTEGER                                  :: KM,k
     
-    REAL*8, PARAMETER                          :: a0 = 999.842594
-    REAL*8, PARAMETER                          :: a1 =   6.793952e-2
-    REAL*8, PARAMETER                          :: a2 =  -9.095290e-3
-    REAL*8, PARAMETER                          :: a3 =   1.001685e-4
-    REAL*8, PARAMETER                          :: a4 =  -1.120083e-6
-    REAL*8, PARAMETER                          :: a5 =   6.536332e-9
+    REAL, PARAMETER                          :: a0 = 999.842594
+    REAL, PARAMETER                          :: a1 =   6.793952e-2
+    REAL, PARAMETER                          :: a2 =  -9.095290e-3
+    REAL, PARAMETER                          :: a3 =   1.001685e-4
+    REAL, PARAMETER                          :: a4 =  -1.120083e-6
+    REAL, PARAMETER                          :: a5 =   6.536332e-9
     
-    REAL*8, PARAMETER                          :: b0 =  8.24493e-1
-    REAL*8, PARAMETER                          :: b1 = -4.0899e-3
-    REAL*8, PARAMETER                          :: b2 =  7.6438e-5
-    REAL*8, PARAMETER                          :: b3 = -8.2467e-7
-    REAL*8, PARAMETER                          :: b4 =  5.3875e-9
+    REAL, PARAMETER                          :: b0 =  8.24493e-1
+    REAL, PARAMETER                          :: b1 = -4.0899e-3
+    REAL, PARAMETER                          :: b2 =  7.6438e-5
+    REAL, PARAMETER                          :: b3 = -8.2467e-7
+    REAL, PARAMETER                          :: b4 =  5.3875e-9
     
-    REAL*8, PARAMETER                          :: c0 = -5.72466e-3
-    REAL*8, PARAMETER                          :: c1 = +1.0227e-4
-    REAL*8, PARAMETER                          :: c2 = -1.6546e-6
+    REAL, PARAMETER                          :: c0 = -5.72466e-3
+    REAL, PARAMETER                          :: c1 = +1.0227e-4
+    REAL, PARAMETER                          :: c2 = -1.6546e-6
     
-    REAL*8, PARAMETER                          :: d0 = 4.8314e-4
+    REAL, PARAMETER                          :: d0 = 4.8314e-4
 
 !    allocate ( dens_temp(KM),dens_zero(KM) ,rho(KM) )
-    REAL*8 t(KM),s(KM)  
-    REAL*8 dens_temp(KM),dens_zero(KM) ,rho(KM) 
+    REAL*4 t(KM),s(KM)  
+    REAL*4 dens_temp(KM),dens_zero(KM) ,rho(KM) 
 
     do k=1,km
        dens_temp(k) = a0+(a1+(a2+(a3+(a4+a5*t(k))*t(k))*t(k))*t(k))*t(k)
